Don't clobber kernel build directory
Currently the kernel build directory is stored in /tmp and thus cannot
be expected to survive reboots.  In addition, after the CM1 kernel is
built, the contents of the directory are overwritten with the CM3 build.
This is undesirable as the build directory may be needed for subsequent
manual builds of the piControl driver.  Fix by using separate build
directories for CM1 and CM3 which are not located in /tmp.

Signed-off-by: Lukas Wunner <lukas@wunner.de>