Synchronize system time from RTC on boot
A bug in /lib/udev/hwclock-set (installed by the util-linux deb package)
prevents synchronization of the system time from the RTC on boot.  As a
workaround, compile the drivers for I2C and RTC into the kernel and
enable CONFIG_RTC_HCTOSYS so that the kernel synchronizes the system
time by itself.  This will also allow synchronization after suspend,
should the Raspberry Pi ever support that.

Signed-off-by: Lukas Wunner <lukas@wunner.de>