gen_bootloader_postinst_preinst: Remove config.txt legacy RTC entry
Wheezy and jessie images enabled the RevPi Core's RTC using an entry
in config.txt.  However the "revpi-core" overlay now also declares the
RTC and with the legacy entry present, we end up with two RTCs in the
devicetree and a bunch of ugly errors on boot.  Avoid by removing the
legacy entry on postinst.

Signed-off-by: Lukas Wunner <lukas@wunner.de>