Update dt-blob.bin on package installation
The devicetree in /boot/dt-blob.bin contains the initial GPIO
configuration set up by the bootloader before it hands over
control to the kernel.

So far we only created this file once on invocation of
revpi-factory-reset by copying the machine-specific
dt-blob.bin from /boot/overlays to /boot.

We now have a case where a machine's dt-blob.bin had to be modified to
adjust to hardware changes.  Namely, revision 06 of the RevPi Connect
dropped a pull-down resistor from GPIO 4 (which controls a GPIO mux),
requiring the bootloader to actively drive the GPIO low.

If an R06 Connect is flashed with an old image, its dt-blob.bin will be
outdated.  However we can ensure that it gets the latest version once
the kernel package is updated:  Amend raspberrypi-kernel.postinst to
copy the new dt-blob.bin from /boot/overlays to /boot.

Signed-off-by: Lukas Wunner <lukas@wunner.de>