build: Make sure required directory "extra" exists
With commit a03872b7d15f directory "extra" was removed from the repository
which now results in an abortion of the build process as soon as the build
script tries to create the file "uname_string" in that directory. So make
sure that directory "extra" is created by the build script if it does not
exist yet.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>