Abandon mkknlimg
The Raspberry Pi Foundations' bootloader used to require a trailer after
the kernel image.  The trailer was added by scripts/mkknlimg.

Starting with rpi-5.3.y, mkknlimg is no longer in the Foundation's
kernel source tree because newer versions of the bootloader can deal
with an unmodified zImage.  Experimentation shows that's been the case
since bootloader version 1.20160506-1.

We shipped 1.20160315-1 with our original wheezy image (still required
the trailer), but 1.20170405-1 with our first jessie image in May 2017
(can deal with an unmodified zImage).

Abandon use of mkknlimg and add a package dependency on the first
bootloader version which no longer requires the trailer.

Link: https://github.com/raspberrypi/linux/issues/3249
Signed-off-by: Lukas Wunner <lukas@wunner.de>