remove redirection to /tmp/out and /tmp/out7
Redirection of kernel build output to /tmp/out and /tmp/out7 removed.
If somebody is interested in analyzing build output it can be redirected
when calling the script.
Removing redirection also fixes the problem when a second user
on the build system tries to build a new kernel and fails because
of missing rights to write to an already existing /tmp/out and /tmp/out7
from a previous build done by another user.

Signed-off-by: Frank Pavlic <f.pavlic@kunbus.com>