Remove elevator=deadline from the cmdline in postinst
Remove the deadline ioscheduler from the cmdline

Since some time now the kernel (at least 4.19) uses the deadline
scheduler by default. With kernel 5.10 the elevator= option is ignored
and a warning is printed in the kernel log:

  Kernel parameter elevator= does not have any effect anymore.
  Please use sysfs to set IO scheduler for individual devices.

We will remove elevator=deadline from /boot/cmdline.txt on the system. The
user is in control of the this file. And we need to make sure this will not
break anything. So we only remove the string if the options before and
after elevator=deadline also match. If the user has changed the cmdline
he needs to remove the elevator= himself.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>