update.sh: check if source folders exists
When building a new kernel package, environment variables must be set
for the linux kernel sources and additional for the picontrol sources.
If the variables are set, there was no check if the folder really
exists. The kernel build would abort in some way. As the build of
picontrol is optional, the build will not check for the source folder of
pictrontrol if it should be built into the package.

Add a check so that the script validates the existence of the folders
if the corresponding environment variable LINUXDIR or PIKERNELMODDIR is
set.

Signed-off-by: Frank Erdrich <f.erdrich@kunbus.com>