Allow for optional arm64 builds
In order to build an optional arm64 kernel with the same toolchain,
the existing build scripts have been modified. With this it is possible
to specify the target architecture via the env var `ARCH`. If `ARCH` is
not specified, the build script assumes `arm` as default (keep previous
behaviour). If `ARCH=arm64` is specified, the build scripts builds only
one kernel (`kernel8.img`) which can be used for CM3 as well as CM4S. It
is now also possible to explicitly override the CROSS_COMPILE var from
the environment. This can be helpfull if the kernel is build on target
architecture.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>