Remove build artifacts with host arch
The current approach with fails on systems which are not x86_64 with
the following error message:

  LD [M]  sound/usb/line6/snd-usb-line6.ko
  LD [M]  sound/usb/line6/snd-usb-toneport.ko
  LD [M]  sound/usb/misc/snd-ua101.ko
  LD [M]  sound/usb/snd-usb-audio.ko
  LD [M]  sound/usb/snd-usbmidi-lib.ko
make[1]: Leaving directory '/home/pi/kernelbakery/kbuild'
rm: missing operand
Try 'rm --help' for more information.

This is caused by the implementation in line 36 of debian/update.sh and
the false assumption that there will be always files matching
'ELF .* x86-64'. On other systems where nothing is found 'xargs rm' is
called with an empty stdin which fails with the error message above.

Fix this by running xargs only on non-empty input and also make sure that
binaries from other buildhost archs are removed.

Reported-by: Simon Han <z.han@kunbus.de>
Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>