Check for blacklist file before working with it
The blacklist file /etc/modprobe.d/blacklist-rtl8192cu.conf is added by our
imagebakery and therefore not part of the official Raspberry Pi OS image. If a
user tries to install our kernel on the vanilla RPi OS this will fail.

Fix this by checking first if the file is present and skip the steps if absent.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>