fix: Use the Date field of the changelog file instead of the timestamp
The option -STimestamp to dpkg-parsechangelog gives a unix timestamp.
But the KBUILD_BUILD_TIMESTAMP is usually not a Unix timestamp, but a
human readable date and time. The time is given in RFC2888 format.

This string is visible with `uname -a`.

Use the Date field instead of the timestamp field.

Fixes: 027226c72091 ("Reproducible build: KBUILD_BUILD_TIMESTAMP")

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>